HV-1667 Allow specific implementations of configuration classes and interfaces
